+++ karma.conf.js
@@ -4,16 +4,13 @@ const webpackConfig = require('./webpack.config');
 module.exports = (config) => {
     config.set({
       files: [
-        // all files ending in "_test"
-        { pattern: 'test/main.ts', watched: false },
-        { pattern: 'test/worker.ts', watched: false },
-        // each file acts as entry point for the webpack configuration
+        { pattern: './test/main.ts', watched: false },
+        { pattern: './test/worker.ts', watched: false },
       ],
       frameworks: ['jasmine'],
       preprocessors: {
-        // add webpack as preprocessor
-        'test/main.ts': ['webpack'],
-        'test/worker.ts': ['webpack'],
+        './test/main.ts': ['webpack'],
+        './test/worker.ts': ['webpack'],
       },
       webpack: {
         ...webpackConfig,
