+++ modules/core/resolver/base-runner.resolver.ts
@@ -1,0 +1,5 @@
+import { Constructor } from "@core/constructor";
+
+export interface IRunnerResolverConfigBase<R extends Constructor> {
+    runners: R[];
+}
