+++ package.json
@@ -25,6 +25,7 @@
     "karma-jasmine": "^2.0.1",
     "karma-webpack": "^4.0.2",
     "ts-loader": "^6.2.1",
+    "tsconfig-paths-webpack-plugin": "^3.2.0",
     "typescript": "^3.7.2",
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.10"
