+++ src/index.ts
@@ -1,5 +1,3 @@
+export { RunnerResolver } from '@modules/promise/runner.resolver';
 export { IRunnerError } from './commands/runner-error';
 export { RunnerErrorCode, RunnerErrorMessages } from './errors/runners-errors';
-export { RunnerResolver } from './resolver';
-
-
