--- src/resolver/node-runner.resolver.ts
@@ -1,81 +1,0 @@
-import { NodeCommand } from "../commands/node-commands";
-import { errorCommandToRunnerError, IRunnerError } from "../commands/runner-error";
-import { Constructor } from "../constructor";
-import { RunnerErrorCode, RunnerErrorMessages } from "../errors/runners-errors";
-import { resolveRunnerBridgeConstructor } from "../runner/bridge-constructor.resolver";
-import { IRunnerBridgeConstructor } from "../runner/runner-bridge";
-import { WorkerBridge } from "../worker-bridge/worker-bridge";
-import { WorkerBridgeBase } from "../worker-bridge/worker-bridge-base";
-import { WorkerBridgeForDev } from "../worker-bridge/worker-bridge-for-dev";
-import { RunnerResolverBase } from "./base-runner.resolver";
-
-export function nodeRunnerResolverMixin<R extends Constructor, T extends new (...args: any[]) => RunnerResolverBase<R>>(runnerResolver: T) {
-    return class NodeRunnerResolver extends runnerResolver {
-        private workerBridges?: WorkerBridgeBase[];
-        private workerIndex = 0;
-        private runnerBridgeConstructors = new Array<IRunnerBridgeConstructor<R>>();
-
-        public async run(): Promise<void> {
-            this.runnerBridgeConstructors = this.config.runners.map(runner => resolveRunnerBridgeConstructor(runner));
-            if (this.config.devMode) {
-                this.workerBridges = [new WorkerBridgeForDev(this as any)];
-                return;
-            }
-            const workerBridgesInits$ = new Array<Promise<WorkerBridge>>();
-            for (let i = 0; i < this.config.totalWorkers; i++) {
-                const bridge = new WorkerBridge({
-                    workerPath: this.config.workerPath,
-                    workerName: `${this.config.namePrefix}${i}`,
-                });
-                workerBridgesInits$.push(bridge.init().then(() => bridge))
-            }
-            const workerBridges = await Promise.all(workerBridgesInits$);
-            this.workerBridges = workerBridges;
-        }
-  
-        public async resolve<RR extends R>(runner: RR, ...args: ConstructorParameters<RR>): Promise<InstanceType<IRunnerBridgeConstructor<RR>>> {
-            const runnerId = this.config.runners.indexOf(runner);
-            if (runnerId < 0) {
-                throw {
-                    error: RunnerErrorMessages.CONSTRUCTOR_NOT_FOUND,
-                    errorCode: RunnerErrorCode.RUNNER_INIT_CONSTRUCTOR_NOT_FOUND,
-                } as IRunnerError;
-            }
-            const workerBridge = this.getNextWorkerBridge();
-            try {
-                await workerBridge.execCommand({
-                    type: NodeCommand.INIT,
-                    instanceId: workerBridge.resolveNewRunnerInstanceId(),
-                    runnerId,
-                    arguments: args,
-                });
-            } catch (error) {
-                throw errorCommandToRunnerError(error);
-            }
-            return new (this.runnerBridgeConstructors[runnerId])(workerBridge, runnerId);
-        }
-
-        /**
-         * Destroy workers for runnable resolver
-         * @param force Destroy by skipping the call the destruction method on the remaining instances
-         */
-        public async destroy(force = false): Promise<void> {
-            if (this.workerBridges) {
-                await Promise.all(this.workerBridges.map(workerBridge => workerBridge.destroy(force)));
-            }
-            this.workerBridges = undefined;
-        }
-
-        private getNextWorkerBridge(): WorkerBridgeBase {
-            if (!this.workerBridges || this.workerBridges.length === 0) {
-                throw new Error('Workers was not started');
-            }
-            const workerIndex = this.workerIndex++;
-            if (this.workerIndex >= this.workerBridges.length) {
-                this.workerIndex = 0;
-            }
-            return this.workerBridges[workerIndex];
-        }
-
-    }
-}
