+++ src/worker-bridge/worker-bridge-for-dev.ts
@@ -1,14 +1,13 @@
 import { INodeCommand } from "../commands/node-commands";
-import { workerRunnerResolverMixin } from "../resolver/worker-runner.resolver";
 import { WorkerBridgeBase } from "./worker-bridge-base";
 
 export class WorkerBridgeForDev extends WorkerBridgeBase {
-    constructor(private runnerResolver: InstanceType<ReturnType<typeof workerRunnerResolverMixin>>) {
-        super();
-        runnerResolver.sendCommand = this.handleWorkerCommand.bind(this);
-    }
+    // constructor(private runnerResolver: InstanceType<ReturnType<typeof workerRunnerResolverMixin>>) {
+    //     super();
+    //     runnerResolver.sendCommand = this.handleWorkerCommand.bind(this);
+    // }
 
     protected sendCommand(command: INodeCommand): void {
-        this.runnerResolver.handleCommand(command);
+        // this.runnerResolver.handleCommand(command);
     }
 }
